functional-tester: move "etcd-runner" to "./cmd"
