+++ tools/functional-tester/runner/root.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package command implements individual etcd-runner commands for the etcd-runner utility.
-package command
+// Package runner implements individual etcd-runner commands for the etcd-runner utility.
+package runner
 
 import (
 	"log"
